change routes to match swagger, put both GET/DELETE functions in the same file respectively
